+++ dsh_sdk/src/protocol_adapters/token/rest_token/mod.rs
@@ -1,0 +1,11 @@
+//! Rest token to be used for fetching [`DataAccessToken`](crate::protocol_adapters::token::data_access_token::DataAccessToken)
+mod claims;
+mod request;
+mod token;
+
+#[doc(inline)]
+pub use claims::{Claims, DatastreamsMqttTokenClaim};
+#[doc(inline)]
+pub use request::RequestRestToken;
+#[doc(inline)]
+pub use token::RestToken;